requirements.txt
streamlit>=1.28.0
supabase>=2.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
requests>=2.28.0
playwright>=1.40.0
